migrator_tool/requirements.txt
google-cloud-aiplatform>=1.38.0
typer>=0.9.0
rich>=13.7.0
pydantic>=2.5.0
pathspec>=0.12.0
jinja2>=3.1.2
